requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.0
pyaudio==0.2.13
pydub==0.25.1
assemblyai==0.17.0
groq==0.4.0
openai==1.3.0
flask-cors==4.0.0
pandas==2.1.0
openpyxl==3.1.2
reportlab==4.0.4
xhtml2pdf==0.2.11
Flask-Mail==0.9.1 